feat: adding tests; adding docs to README

+++ src/index.ts
@@ -12,8 +12,8 @@ const app = express();
 
 app.use(express.json());
 
-app.get("/api/generate-random-number", async (req, res) => {
-  return res.json(Math.floor(Math.random() * 50));
+app.get("/api/generate-random-number", async (_, res) => {
+  res.json(Math.floor(Math.random() * 50));
 });
 
 // Get request so it can be tested in the browser
@@ -34,7 +34,7 @@ app.get("/api/entity/:id/update-number", async (req, res) => {
       select: { id: true, number: true },
       where: { id: Number(id) },
       update: { number: (existingEntity?.number || 0) + Number(randomNumber) },
-      create: { id: Number(id), number: 1 },
+      create: { number: 1 },
     });
 
     res.json(newEntity);
@@ -43,30 +43,47 @@ app.get("/api/entity/:id/update-number", async (req, res) => {
   }
 });
 
-app.listen(3000, () =>
+export const server = app.listen(3000, () =>
   console.info("🚀 Server ready at: http://localhost:3000")
 );
 
-cron.schedule("30 * * * * *", async () => {
-  console.info("🕒 Running a cron clean up");
+export const cleanup = cron.schedule(
+  "30 * * * * *",
+  async () => {
+    console.info("🕒 Running a cron clean-up");
 
-  try {
-    const foundEntities = await prisma.entity.findMany({
-      where: {
-        number: {
-          gte: 10,
+    try {
+      const foundEntities = await prisma.entity.findMany({
+        select: {
+          id: true,
+          number: true,
         },
-      },
-    });
 
-    foundEntities.forEach(async (entity) => {
-      await prisma.entity.delete({
         where: {
-          id: entity.id,
+          number: {
+            gte: 10,
+          },
         },
       });
-    });
-  } catch (error) {
-    console.error(error);
-  }
-});
+
+      await Promise.all(
+        foundEntities.map(async (entity) => {
+          if (entity.number % 2 === 0 || entity.number % 10 === 0) {
+            await prisma.entity.delete({
+              where: {
+                id: entity.id,
+              },
+            });
+          }
+        })
+      );
+    } catch (error) {
+      console.error(error);
+    }
+  },
+  { scheduled: false }
+);
+
+if (process.env.NODE_ENV !== "test") {
+  cleanup.start();
+}
